working on restructuring main
